Rejigger tests
