+++ nnll.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 2.4
 Name: nnll
-Version: 0.1.dev198+g46b57b8.d20250322
+Version: 0.1.dev199+g9af0ccc.d20250322
 Summary: Neural Network Link Library : A comprehensive modular toolkit for Diffusion and Large Language Model inference processes.
 Author-email: darkshapes <91800957+exdysa@users.noreply.github.com>
 License: #// SPDX-License-Identifier: blessing
@@ -100,12 +100,6 @@ Requires-Dist: nnll[nnll_63]; extra == "all"
 Requires-Dist: nnll[nnll_64]; extra == "all"
 Provides-Extra: dev
 Requires-Dist: nnll[all]; extra == "dev"
-Requires-Dist: pytest; extra == "dev"
-Requires-Dist: ruff; extra == "dev"
-Requires-Dist: textual-dev; extra == "dev"
-Requires-Dist: pytest-asyncio; extra == "dev"
-Requires-Dist: pytest-tornasync; extra == "dev"
-Requires-Dist: pytest-trio; extra == "dev"
 Dynamic: license-file
 
 <div align="center">
@@ -119,6 +113,7 @@ Dynamic: license-file
 
 Emphasis is on refining and optimizing code for extracting and classifying metadata, ML pipeline preparation, GPU configuration, consumer-grade system optimization, and a comprehensive suite of direct and indirect generative AI preparations for 🧨Diffusers, 🤗Transformers, 🦙Ollama and 🌀DSPy frameworks.
 
+Thank you for looking forward to it.
 
 <br>
 
@@ -140,7 +135,7 @@ Some modules are full scripts and can be run from command line. These are writte
 
 Each module contains 1-5 functions or 1-2 classes and its own test routines. There are multiple ways to integrate nnll into a project (sorted by level of involvement)
 
-- *Recommended* : Add the project as a dependency including only modules that are needed with `"nnll[04,16]" @ git+https://github.com/darkshapes/nnll`
+- *Recommended* : Add the project as a dependency including only modules that are needed with `"nnll[nnll_04,nnll_16]" @ git+https://github.com/darkshapes/nnll`
 - Install the entire project as a dependency via `nnll @ git+https://github.com/darkshapes/nnll`
 - Basic clone or fork of the project
 -  Use a [submodule](https://github.blog/open-source/git/working-with-submodules/)
@@ -186,11 +181,11 @@ Each module contains 1-5 functions or 1-2 classes and its own test routines. The
 > ```
 > or for select packages
 > ```
-> pip install -e "nnll[33,56]"
+> pip install -e "nnll[nnll_33,nnll_56]"
 > ```
-or for all packages :
+> or for all packages :
 > ```
-> pip install -e 'nnll\[dev\]'
+> pip install -e "nnll[dev]"
 >```
 
 ##### Done.
