+++ nnll.egg-info/SOURCES.txt
@@ -135,7 +135,6 @@ tests/__init__.py
 tests/test.py
 tests/test_combo_history.db
 tests/test_img.png
-tests/textual.log
 tests/10/test_init_exit.py
 tests/10/test_model_register.py
 tests/10/test_orientation.py
