+++ nnll_10/package/__main__.py
@@ -19,7 +19,7 @@ class Combo(App):
     CSS_PATH = "combo.tcss"
     BINDINGS = [Binding("escape", "safe_exit", "◼︎ / ⏏︎")]  # Cancel response
 
-    safety: reactive[int] = reactive(1)
+    safety: reactive[int] = reactive(2)
 
     def on_mount(self) -> None:
         """Draw screen"""
@@ -34,7 +34,6 @@ class Combo(App):
         if event.key not in ["escape", "ctrl+left_square_brace"]:
             self.safety += 1
         else:
-            # event.prevent_default()
             self.safety -= 1
             if self.safety < 0:
                 event.prevent_default()
